+++ examples/graceful_shutdown_web_sockets/src/main.rs
@@ -176,10 +176,10 @@ async fn main() -> anyhow::Result<()> {
                     // Larger timeouts to demonstrate rapid graceful shutdown
                     static CONFIG: picoserve::Config =
                         picoserve::Config::new(picoserve::Timeouts {
-                            start_read_request: Some(Duration::from_secs(10)),
-                            persistent_start_read_request: Some(Duration::from_secs(10)),
-                            read_request: Some(Duration::from_secs(1)),
-                            write: Some(Duration::from_secs(1)),
+                            start_read_request: Duration::from_secs(10),
+                            persistent_start_read_request: Duration::from_secs(10),
+                            read_request: Duration::from_secs(1),
+                            write: Duration::from_secs(1),
                         })
                         .keep_connection_alive();